Fix for spurious delete of default toks.
Problem scenario:
    fu   xxx   f-   fi
When cleanup considers f- and subsequent fi, it concludes that the f- is
superseded (because there's nothing that would be "hlable" by f-).
Problem: Deleting the f- allows the whitespace between f- and fi to appear
underlined!!! In other words, we can't test for hlable considering only the
potentially superseded token; we must consider the effect of a preceding token
that could "bleed through". The correct action here would probably be to leave
the f- alone; if it had been a non-default tok (e.g., bold), it would probably
be best to replace the fb with f-, on the grounds that it's more aesthetically
pleasing to leave behind a default tok when that token's sole purpose is to
prevent bleed through...