Fixed bug breaking highlight of small regions.
Bug was caused by Vmap_cleanup's failure to take token's action into account
when deciding between inc/exc args. Also note that because the function mapping
action to inc/exc is different for start and end, it's necessary to pass the arg
as length 2 array. Was thinking at first that the problem here was large, but
thought it over and realized that because I'm dealing with an as-yet unmodified
buffer, it's really not.

Also, put in stopgap logic that implements something like the final
bleed-through supersedence logic, though the actual logic is going to require a
separate function, or a modified version of Contains_hlable. (TODO)