Added basic support for highlight/delete operators
Note: Support incomplete: in particular, need to modify logic for
adjustment of '[ and '] following and operator application. This is a
bit tricky because the positions I pass into Operate_region() are not
the ones corresponding to the real '[ and '], so after operation has
completed, I currently have no idea how the real marks have been
impacted. Need to add logic to Vmap_apply_changes to support adjustment
of arbitrary positions.