Reworked Parse_fmt_clr_transformer to permit slightly more flexible format: in particular, don't require the strict Perl regex flag-like syntax: i.e., 	fxxx-yyy ...but permit + and - components to be interspersed freely, provided there are no ambiguities: e.g., 	fu-i+b-s   OK 	fu-iu+b    ERROR! Ambiguity with `u' Note: Currently, this reworked function errs on the side of not permitting things that could cause confusion: e.g., multiple components of the same type (fmt/clr/bgc) or empty component (no type), but I've left some configuration flags in the function that permit toggling various aspects of the behavior as necessary to test out usability. I'm thinking this could be useful as I begin to get feedback from alpha users.
Note: The old function is left as <...>_obsolete for now, but will
delete or move to code graveyard after commit.

Change: Under the new strategy (with current configuration flags),
hyphen in auto map fmt specs is not special: i.e., the following are all
do-nothings:
	f
	f+
	f-
Note that for manual maps and auto map clr/bgc specs, `-' still has its
special meaning.
TODO: If users are bothered by this, the config flags make it easy to
change back to f- being a special "no fmt" designator.

Reworked Translate_fmt_clr_spec to make it a bit more consistent with
Parse_fmt_clr_transformer: in particular, it now allows defaults to be
specified both with and without `-'. Users shouldn't mind this, as they
can continue typing the `-' if they like, but don't have to...