Completely reworked logic pertaining to the Visual shortcuts.
The new logic reflects a better understanding of Vim's implementation of
select-mode maps, as described in Vim doc section 'Select-mode-mapping'.
The bottom line is that there are reasons to use either vmap or xmap for
shortcuts, but never smap.
Also implemented 2 new flag options:
	txtfmtShortcutsWorkInSelect
	txtfmtShortcutsStayInVisual
More testing and cleanup is needed before release, but the branch is
probably ready for users who don't mind beta testing.